Fully Complete Pick Up Weapon Mechanics
You can now Pick Up a Weapon In the Game and The Ownership of that weapon is transferred to the client that holds it.

The Transform of the weapon Also updates accordingly, and sends the stream info to the server and back to all other clients.

The ownership is also updated to All Buffered Players, so any one who joins will receive the updates.